Fix declaration merging for express
