--- server/express.d.ts
@@ -1,9 +1,0 @@
-import type { AuthenticationStatus } from '#server/routes/accounts';
-
-declare global {
-  namespace Express {
-    export interface Request {
-      authentication?: AuthenticationStatus;
-    }
-  }
-}
